+++ src/messages/data.cpp
@@ -57,8 +57,8 @@ uint64_t data::process_payload(std::istream& input,uint64_t size) {
     }
     
     if (new_size > m_payload.capacity()) {
-        m_payload.reserve(std::max<uint64_t>(
-            new_size, static_cast<uint64_t>(2*m_payload.capacity())
+        m_payload.reserve(std::max<size_t>(
+            new_size, static_cast<size_t>(2*m_payload.capacity())
         ));
     }
